+++ apps/admin-web/next.config.js
@@ -12,9 +12,9 @@ const nextConfig = {
   },
   reactStrictMode: true,
   output: 'standalone',
+  transpilePackages: ['@cgr/course-utils', '@cgr/codegen'],
   experimental: {
     outputFileTracingRoot: path.join(__dirname, '../../'),
-    transpilePackages: ['@cgr/course-utils', '@cgr/codegen'],
   },
 }