+++ apps/admin-web/src/env/index.ts
@@ -1,0 +1,4 @@
+export const BACKEND_URI = process.env.NEXT_PUBLIC_BACKEND_URL || ''
+export const ENVIRONMENT = process.env.NEXT_PUBLIC_ENVIRONMENT || 'local'
+
+export const ADMIN_ACCESS_TOKEN = process.env.NEXT_PUBLIC_ADMIN_ACCESS_TOKEN || ''