+++ apps/web/.eslintrc.js
@@ -1,0 +1,1 @@
+module.exports = require('@cgr/project-config/eslint/next')