+++ apps/web/CHANGELOG.md
@@ -1,5 +1,15 @@
 # web
 
+## 1.2.0-beta.0
+
+### Minor Changes
+
+- 4af83ff9: update eslint config, fixed eslint error
+
+### Patch Changes
+
+- 0f2fab86: update prettier pipeline and import order
+
 ## 1.1.1
 
 ### Patch Changes